Práctica_obligatoria_2_Zamorano_Marcos/python-api/requirements.txt
Flask==3.0.3
requests==2.32.3
mysql-connector-python==8.4.0